PB/course_project/requirements.txt
Pillow==9.3.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
haversine==2.7.0